Updating existing integration tests for static typing (#522)

+++ src/commonMain/kotlin/indexer/codegen/IndexerAssetPositionResponseObject.kt
@@ -28,5 +28,5 @@ data class IndexerAssetPositionResponseObject(
     val side: IndexerPositionSide? = null,
     val size: kotlin.String? = null,
     val assetId: kotlin.String? = null,
-    val subaccountNumber: kotlin.Double? = null
+    val subaccountNumber: kotlin.Int? = null
 )
